Gui: Move value conversions into their own headers
Utilities.h carried both the Base trait and convertTo specialisations for
Coin, Qt and style parameter types, and the unrelated Gui helpers. Every
consumer parsed QBrush, the gradient headers and three StyleParameters
headers to reach them, however little of that it used.

The conversions move to Gui/Conversion/{Coin,Qt,StyleParameters}.h and each
consumer now includes the one it needs. Five includes turned out to be unused
and are gone; three files that only wanted convertTo now say so directly.
