+++ src/Gui/Dialogs/DlgThemeEditor.cpp
@@ -28,7 +28,8 @@
 #include "ui_DlgThemeEditor.h"
 #include "BitmapFactory.h"
 
-#include <Utilities.h>
+#include <Conversion/Qt.h>
+#include <Conversion/StyleParameters.h>
 #include <Base/ServiceProvider.h>
 #include <Base/Tools.h>
