+++ src/Mod/MeshPart/Gui/CurveOnMesh.cpp
@@ -54,7 +54,7 @@
 #include <Base/Converter.h>
 #include <Gui/Document.h>
 #include <Gui/MainWindow.h>
-#include <Gui/Utilities.h>
+#include <Gui/Conversion/Coin.h>
 #include <Gui/View3DInventor.h>
 #include <Gui/View3DInventorViewer.h>
 #include <Mod/Mesh/App/Core/Algorithm.h>
