+++ src/Mod/Part/Gui/ViewProviderPreviewExtension.cpp
@@ -35,7 +35,7 @@
 #include <App/Document.h>
 #include <App/ExtensionPython.h>
 #include <Base/ServiceProvider.h>
-#include <Gui/Utilities.h>
+#include <Gui/Conversion/Coin.h>
 #include <Gui/Inventor/So3DAnnotation.h>
 #include <Mod/Part/App/PreviewExtension.h>
 #include <Mod/Part/App/Tools.h>
