+++ src/Mod/PartDesign/Gui/ViewProvider.cpp
@@ -40,7 +40,7 @@
 #include <Gui/Selection/SoFCUnifiedSelection.h>
 #include <Gui/Inventor/So3DAnnotation.h>
 #include <Gui/MainWindow.h>
-#include <Gui/Utilities.h>
+
 #include <Mod/PartDesign/App/Body.h>
 #include <Mod/PartDesign/App/FeatureAddSub.h>
 #include <Mod/Part/Gui/ViewProvider.h>
