+++ src/Mod/PartDesign/Gui/ViewProviderHole.cpp
@@ -57,7 +57,7 @@
 #include <Base/Placement.h>
 #include <Base/Tools.h>
 #include <App/Property.h>
-#include <Utilities.h>
+#include <Conversion/Coin.h>
 
 #include <Inventor/nodes/SoClipPlane.h>
 #include <Inventor/nodes/SoCoordinate3.h>
