+++ src/Mod/PartDesign/Gui/ViewProviderSketchBased.cpp
@@ -27,7 +27,6 @@
 #include <App/Document.h>
 #include <Base/ServiceProvider.h>
 #include <Gui/Application.h>
-#include <Gui/Utilities.h>
 #include <Mod/Sketcher/App/SketchObject.h>
 #include <Mod/PartDesign/App/FeatureSketchBased.h>
